Add mailbox x-guid handling.
Some mail clients use a mailbox x-guid to identify a mailbox uniquely
rather than by name (SoGo for one).  Without this, some clients will
repeatedly query for the x-guid and have problems with mailbox sync.

This patch adds this but relies on the uuid type being handled by
the PostgreSQL components.  A prior patch includes that.